+++ src/components/index.ts
@@ -9,7 +9,7 @@ export { SubHeader } from './header/SubHeader'
 export * from './icons'
 export * from './inputGroup'
 export { default as Label } from './label/Label'
-export * from './modal/Modal'
+export { default as Modal } from './modal/Modal'
 export * from './portal/Portal'
 export * from './profile/Profile'
 export * from './slider/TimeSlider'
