+++ src/features/clinic/ui/AlarmBottomSheet.tsx
@@ -4,9 +4,8 @@ import { useFormContext } from 'react-hook-form'
 
 import { BottomSheet, Label, SubHeader } from '@/components'
 import { ALARM_TIME } from '@/constants'
+import { useSelectAlarmTime } from '@/features'
 import type { ClinicBottomSheetType } from '@/types'
-
-import { useSelectAlarmTime } from '../hook/useSelectAlarmTime'
 
 export const AlarmBottomSheet = ({ section, isShowing, onClickScrim }: ClinicBottomSheetType) => {
   const { setValue } = useFormContext()
