+++ src/features/index.ts
@@ -8,6 +8,15 @@ export * from './sign-up/TosStep'
 export * from './sign-up/UserInfoStep'
 
 // clinic
+export * from './clinic/hook/useClinicFormBottomSheet'
+export * from './clinic/hook/useFormattedVisitDate'
+export * from './clinic/hook/useSelectAlarmTime'
+export * from './clinic/hook/useTagToggle'
+export * from './clinic/service/useAddClinicFormMethod'
+export * from './clinic/service/useClinicDeleteModal'
+export * from './clinic/service/useEditClinicFormMethod'
+export * from './clinic/service/useInsertedClinicForm'
+export * from './clinic/service/useScheduleList'
 export * from './clinic/ui/AlarmBottomSheet'
 export * from './clinic/ui/ClinicClientPage'
 export * from './clinic/ui/ClinicForm'
